backend/requirements.txt

Django>=4.0
djangorestframework
django-cors-headers
google-generativeai
google-cloud-speech
google-cloud-texttospeech
google-cloud-translate
requests
python-dotenv
